requirements.txt
alembic==1.13.1
astroid==3.2.2
blinker==1.8.2
click==8.1.7
dill==0.3.8
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomlkit==0.12.5
typing_extensions==4.11.0
Werkzeug==3.0.3
WTForms==3.1.2
